# Lightweight observable evaluation