Transitions from replanning expressions on each invocation to generating
stateful metadata in a concurrency friendly manner. Offers a 2-4x speedup
over the previous method.

Before:

```
BenchmarkEvalOptions/optimize-12         	 2753293	       571.3 ns/op	      32 B/op	       2 allocs/op
BenchmarkEvalOptions/track-state-12      	  170517	      6829 ns/op	    1729 B/op	      42 allocs/op
BenchmarkEvalOptions/exhaustive-eval-12  	  127220	      8569 ns/op	    1785 B/op	      44 allocs/op
```

After:

```
BenchmarkEvalOptions/optimize-12         	 3305553	       526.5 ns/op	      32 B/op	       2 allocs/op
BenchmarkEvalOptions/track-state-12      	  595286	      1957 ns/op	     368 B/op	       8 allocs/op
BenchmarkEvalOptions/exhaustive-eval-12  	  578535	      2251 ns/op	     368 B/op	       8 allocs/op
```