@jpbetz 2025-04-03
LGTM

I reviewed the change to the planner and that looks how I'd hoped. Bench numbers look really good.